[syncd] Add the possibility to override watchdogWarnTimeSpan timeout with command line option.
If the initialization time of the platform takes more than watchdogWarnTimeSpan
timeout which is by default 30 seconds syncd will print an error message in the syslog.
This is confusing because the system continues to run and work properly.

This commit adds a possibility to override watchdogWarnTimeSpan timeout with the command line parameter.
This allows overriding the default value for the platforms whose initialization takes more than 30 seconds.